fix: improve mobile tierlist editing
